ai-ml/use-cases/11_personalized_communication_nudging/requirements.txt
pandas>=1.4
psycopg2-binary>=2.9
pyyaml>=6.0
Flask>=2.0
scikit-learn>=1.3.0
numpy>=1.24.0

